Enable image cache

+++ SamsClub/AutoLayoutService.swift
@@ -8,18 +8,18 @@
 
 import UIKit
 class AutoLayoutService {
-    static func pin(target view: UIView, to container: UIView, top: Double? = nil, right: Double? = nil, bottom: Double? = nil, left: Double? = nil) {
+    static func pin(target view: UIView, to container: UIView, top: CGFloat? = nil, right: CGFloat? = nil, bottom: CGFloat? = nil, left: CGFloat? = nil) {
         if let top = top {
-            view.topAnchor.constraint(equalTo: container.topAnchor, constant: CGFloat(top)).isActive = true
+            view.topAnchor.constraint(equalTo: container.topAnchor, constant: top).isActive = true
         }
         if let right = right {
-            view.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: CGFloat(right)).isActive = true
+            view.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: right).isActive = true
         }
         if let bottom = bottom {
-            view.bottomAnchor.constraint(equalTo: container.bottomAnchor, constant: CGFloat(bottom)).isActive = true
+            view.bottomAnchor.constraint(equalTo: container.bottomAnchor, constant: bottom).isActive = true
         }
         if let left = left {
-            view.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: CGFloat(left)).isActive = true
+            view.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: left).isActive = true
         }
     }
 }
